requirements.txt
flask==3.0.0
google-api-python-client==2.108.0
plotly==5.18.0
pandas
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0
transformers
torch
numpy==1.26.2 
